requirements/development.txt
-r base.txt

django-debug-toolbar>=1.3.0
django-extensions>=1.5.2
yanc==0.3.3
ipython>=3.1.0
ipdb>=0.8
nose==1.3.6
django-nose==1.4
coverage==3.7.1
factory-boy==2.5.2
